skillbridge_backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
django-channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
mysqlclient==2.2.0
Pillow==10.0.0
celery==5.3.0
redis==5.0.0
requests==2.31.0
razorpay==1.3.0
gunicorn==21.2.0
